+++ app/src/main/res/layout/fragment_weather.xml
@@ -5,8 +5,5 @@
     android:layout_height="match_parent"
     tools:context=".ui.WeatherFragment">
 
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/recyclerView"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent" />
+
 </FrameLayout>
